# Faster join