--- a/toolz/itertoolz.py
+++ b/toolz/itertoolz.py
@@ -875,14 +875,14 @@ def join(leftkey, leftseq, rightkey, rightseq,
 
     d = groupby(leftkey, leftseq)
 
-    if (left_default == no_default) and (right_default == no_default):
+    if left_default == no_default and right_default == no_default:
         # Inner Join
         for item in rightseq:
             key = rightkey(item)
             if key in d:
                 for left_match in d[key]:
                     yield (left_match, item)
-    elif (left_default != no_default) and (right_default == no_default):
+    elif left_default != no_default and right_default == no_default:
         # Right Join
         for item in rightseq:
             key = rightkey(item)
@@ -891,7 +891,7 @@ def join(leftkey, leftseq, rightkey, rightseq,
                     yield (left_match, item)
             else:
                 yield (left_default, item)
-    elif (right_default != no_default):
+    elif right_default != no_default:
         seen_keys = set()
 
         if left_default == no_default: