@groutr 2018-11-07
Working with 4 variations of join.
`join1` is the original implementation of join.
`join2` is changing the try/except to an if statement
`join3` is adding the special case for inner/right joins
`join4` detects each of the 4 join cases and defines a generator purpose-built for that join and returns the result of that generator. (Inspired by the cytoolz implementation).

Implementations here: https://gist.github.com/groutr/19b9f6ba71686e683a30af3eecd3f895

```
In [15]: test(1000, 1000000) # len(leftseq) = 1000, len(rightseq)=1000000
Inner Join (20% overlap)
665 ms ± 2.14 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
424 ms ± 1.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
219 ms ± 4.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
213 ms ± 1.79 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Right Join (20% overlap)
775 ms ± 3.09 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
543 ms ± 9.17 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
327 ms ± 640 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
328 ms ± 5.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Left Join (20% overlap)
665 ms ± 7.36 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
427 ms ± 6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
429 ms ± 6.73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
426 ms ± 4.28 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Full Join (20% overlap)
775 ms ± 4.06 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
550 ms ± 19.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
577 ms ± 84 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
539 ms ± 5.21 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Inner Join (50% overlap)
662 ms ± 3.18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
430 ms ± 12.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
216 ms ± 461 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
213 ms ± 919 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
Right Join (50% overlap)
788 ms ± 13.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
537 ms ± 2.47 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
327 ms ± 2.81 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
325 ms ± 1.79 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Left Join (50% overlap)
741 ms ± 24.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
448 ms ± 9.51 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
474 ms ± 9.92 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
433 ms ± 9.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Full Join (50% overlap)
802 ms ± 10.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
567 ms ± 20.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
598 ms ± 21.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
559 ms ± 13.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Inner Join (99% overlap)
691 ms ± 18.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
444 ms ± 27.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
231 ms ± 5.46 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
226 ms ± 4.87 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Right Join (99% overlap)
826 ms ± 25.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
549 ms ± 11 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
336 ms ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
354 ms ± 12.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Left Join (99% overlap)
682 ms ± 12.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
453 ms ± 14.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
451 ms ± 25.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
432 ms ± 7.76 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Full Join (99% overlap)
825 ms ± 20.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
570 ms ± 27.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
561 ms ± 24 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
566 ms ± 20.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```