@eriknw 2018-11-08
This looks great, thanks again for all your attention recently @groutr!  I'm back home now and more or less caught up with things (it's amazing how disruptive >18000 miles of road trips are), so I'll be giving `toolz` much more attention than I have been.