@groutr 2018-11-08
@eriknw Glad to hear you're back safe.  I should probably clean up these commits and open a new PR.  The commit history in this PR is somewhat messy.

I would like to know your thoughts on the line between clear/easy-to-read and performance.  I wasn't sure if I was crossing that line by treating individual cases separately (and having some duplicated code).  I settled on a variation of `join4`. I think I like having each case wrapped in its own generator, even though it duplicates some code.  When profiling/debugging, it becomes very clear what join is happening because all the action happens in a function named after the type of join. We could move the generators out of the `join` function so they are defined at import time and can be called explicitly, but still not part of the public API.