@eriknw 2018-11-13
I think this looks pretty good.

Generally speaking, `cytoolz` is where the dirtiest, ugliest optimizations go, but we still care about performance of `toolz`.  We care about readability as well, so I prefer this version over #429, because having all the code in a single function body is much more readable imho (b/c one can do `toolz.join??` in IPython).  I actually don't mind having these four cases split out explicitly.