@groutr 2018-12-18
@eriknw I think is ready for another review.