@eriknw 2019-06-22
LGTM.  Thanks again @groutr, and sorry again for the long delay.  Pretty nice performance gain here!